Restructure app layout
