+++ hub-view/src/views/topBar.js
@@ -1,0 +1,12 @@
+import React from "react"
+import Logo from "./logo";
+import SearchBar from "./searchBar";
+
+export default function TopBar(props) {
+  return (
+    <div className="topBar-container">
+        <Logo />
+        <SearchBar sendUserSearchTerm={props.sendUserSearchTerm} sendUrl={props.sendUrl} userSearchTerm={props.userSearchTerm}/>
+    </div> 
+  );
+}
